core/requirements.txt
flask
flask-restx
flask-SQLAlchemy
psycopg2-binary
google-generativeai
google-api-python-client
google-auth
pandas
pytest
black
ollama
sqlalchemy
numpy
